Created entity class for auth
