+++ app/src/androidTest/java/com/skinconnect/userapps/ui/auth/login/LoginFragmentTest.kt
@@ -10,8 +10,6 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.skinconnect.userapps.R
 import com.skinconnect.userapps.data.remote.ApiConfig
 import com.skinconnect.userapps.data.repository.EspressoIdlingResource
-import com.skinconnect.userapps.util.JsonConverter
-import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.MockWebServer
 import org.hamcrest.CoreMatchers.startsWith
 import org.junit.After
